Optimization: Create function `basic_outbuff::write`
It significantly improved the performance of string
arguments,  accorging to benchmarks.
Probably this is because it is inline while the
function it replaces ( `strf::write` ) could not
be, since it contains a loop.

`write` calls a virtual function ( `do_write` ) when
the string length is greater than `size()`, but such
condition is false in most cases.