Replace class invalid_seq_notifer by transcoding_error_notifer
This enables a more detailed error notification.
The new class differentiates invalid sequences ( which has to do
with an invalid input ) from unsupported codepoint ( which has to
do with a limitation of the destination encoding ).
Also, the member functions accept parameters to inform
the invalid sequence or the unsupported codepoint.