Remove surrogate_policy facet.
Because I realized that the calculation of the width of a
transcoded string was incorrect when using
surrogate_policy::lax.

The correct approach in this case would be to transcode the
output ( not the input ) string to UTF-32 using
surrogate_policy::strict ( not lax ), then calcutating the
width of this.

This is a little bit complicated, and since I guess it's not
an important feature, I decided to remove it instead of fix it.